zed_env_vars: Split into helper crate and Zed-specific
